requirements_httpie.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Cython==0.29.30
distlib==0.3.4
Django==4.0.5
django-crispy-forms==1.14.0
django-taggit==3.0.0
django-tinymce==3.4.0
djangorestframework==3.13.1
filelock==3.6.0
Flask==2.1.2
fonttools==4.32.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.0
pipenv==2022.4.21
platformdirs==2.5.2
pygame==2.1.2
pyparsing==3.0.8
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
sqlparse==0.4.2
threadpoolctl==3.1.0
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
Werkzeug==2.1.2
